adjust project for multi module
