+++ .gitignore
@@ -1,6 +1,15 @@
+# Env vars
 .env
 .envrc
 
+# Backend specific mumbojumbo
+*.iml
+*.log
+target/
+bash.exe.stackdump
+users.txt
+.gradle
+build/
 .idea
 **/.DS_Store
 
@@ -10,3 +19,12 @@ npm-debug.log*
 
 # Logs and task output from turborepo
 .turbo
+
+# Ignore Gradle GUI config
+gradle-app.setting
+
+# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
+!gradle-wrapper.jar
+
+# Cache of project
+.gradletasknamecache
