+++ build.gradle.kts
@@ -1,0 +1,11 @@
+
+plugins {
+    val kotlinVersion = "1.6.10"
+    kotlin("jvm") version kotlinVersion apply false
+    kotlin("plugin.serialization") version kotlinVersion apply false
+    id("org.jlleitschuh.gradle.ktlint") version "10.2.1" apply false
+}
+
+dependencies {
+
+}
