--- mulighetsrommet-api/settings.gradle.kts
@@ -1,5 +1,0 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
-rootProject.name = "no.nav.mulighetsrommet.api"
