+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/domain/TiltaksgjennomforingTable.kt
@@ -1,25 +1,11 @@
 package no.nav.mulighetsrommet.api.domain
 
-import kotlinx.serialization.Serializable
 import no.nav.mulighetsrommet.api.database.PGEnum
-import no.nav.mulighetsrommet.api.utils.DateSerializer
+import no.nav.mulighetsrommet.domain.Tiltakskode
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.`java-time`.datetime
 import java.time.LocalDateTime
-
-@Serializable
-data class Tiltaksgjennomforing(
-    val id: Int? = 0,
-    val tittel: String,
-    val beskrivelse: String,
-    val tiltakskode: Tiltakskode,
-    val tiltaksnummer: Int,
-    @Serializable(with = DateSerializer::class)
-    val fraDato: LocalDateTime? = null,
-    @Serializable(with = DateSerializer::class)
-    val tilDato: LocalDateTime? = null
-)
 
 object TiltaksgjennomforingTable : IntIdTable() {
     val tiltakskode: Column<Tiltakskode> = customEnumeration("tiltakskode", "tiltakskode", { value -> Tiltakskode.valueOf(value as String) }, { PGEnum("tiltakskode", it) }).references(TiltakstypeTable.tiltakskode)
