--- mulighetsrommet-domain/lib/src/test/kotlin/mulighetsrommet/domain/LibraryTest.kt
@@ -1,14 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package mulighetsrommet.domain
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class LibraryTest {
-    @Test fun someLibraryMethodReturnsTrue() {
-        val classUnderTest = Library()
-        assertTrue(classUnderTest.someLibraryMethod(), "someLibraryMethod should return 'true'")
-    }
-}
