+++ mulighetsrommet-domain/src/main/kotlin/no.nav.mulighetsrommet.domain/Tiltaksgjennomforing.kt
@@ -1,7 +1,6 @@
-package no.nav.mulighetsrommet.kafka.domain
+package no.nav.mulighetsrommet.domain
 
 import kotlinx.serialization.Serializable
-import no.nav.mulighetsrommet.kafka.utils.DateSerializer
 import java.time.LocalDateTime
 
 @Serializable
