+++ mulighetsrommet-domain/src/main/kotlin/no.nav.mulighetsrommet.domain/Tiltakstype.kt
@@ -1,11 +1,6 @@
-package no.nav.mulighetsrommet.api.domain
+package no.nav.mulighetsrommet.domain
 
 import kotlinx.serialization.Serializable
-import no.nav.mulighetsrommet.api.database.PGEnum
-import no.nav.mulighetsrommet.api.utils.DateSerializer
-import org.jetbrains.exposed.dao.id.IntIdTable
-import org.jetbrains.exposed.sql.Column
-import org.jetbrains.exposed.sql.`java-time`.datetime
 import java.time.LocalDateTime
 
 enum class Tiltakskode {
@@ -30,16 +25,3 @@ data class Tiltakstype(
     @Serializable(with = DateSerializer::class)
     val updatedAt: LocalDateTime? = null
 )
-
-object TiltakstypeTable : IntIdTable() {
-    val navn: Column<String> = text("navn")
-    val innsatsgruppeId: Column<Int> = integer("innsatsgruppe_id").references(InnsatsgruppeTable.id)
-    val sanityId: Column<Int?> = integer("sanity_id").nullable()
-    val tiltakskode: Column<Tiltakskode> = customEnumeration("tiltakskode", "tiltakskode", { value -> Tiltakskode.valueOf(value as String) }, { PGEnum("tiltakskode", it) })
-    val fraDato: Column<LocalDateTime?> = datetime("dato_fra").nullable()
-    val tilDato: Column<LocalDateTime?> = datetime("dato_til").nullable()
-    val createdBy: Column<String?> = text("created_by").nullable()
-    val createdAt: Column<LocalDateTime> = datetime("created_at")
-    val updatedBy: Column<String?> = text("updated_by").nullable()
-    val updatedAt: Column<LocalDateTime> = datetime("updated_at")
-}
