--- mulighetsrommet-kafka/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "no.nav.mulighetsrommet.kafka"
