+++ settings.gradle.kts
@@ -1,0 +1,4 @@
+rootProject.name = "mulighetsrommet"
+include("mulighetsrommet-api")
+include("mulighetsrommet-domain")
+include("mulighetsrommet-kafka")
